tools/cli-interfaces/open-interpreter/requirements.txt
# Open Interpreter AI-Server Integration Requirements
open-interpreter>=0.2.0
aiohttp>=3.8.0
requests>=2.28.0
python-dotenv>=0.19.0